+++ tests/test_inspect.py
@@ -1,0 +1,3 @@
+def test_inspect_func():
+    print(__package__)
+    breakpoint()