+++ HISTORY.rst
@@ -5,7 +5,7 @@ History
 8.0.0b2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- `Issue #164 <https://github.com/nazrulworld/fhir.resources/issues/164>`_ ``get_fhir_model_class`` is now available and can be imported from base package. For example ``from fhir.resources import get_fhir_model_class`` or ``from fhir.resources.STU3 import get_fhir_model_class``
 
 
 8.0.0b1 (2024-08-05)